==> 01b_severity_score.ipynb <==
# 1b. Data Cleaning: Adding a Severity Score Based on the "Reason for Retraction" Column 


## Introduction


This notebooks adds an additional column to the clean data set produced by Notebook 1a. This additional column contains a **severity score** that indicates how severe the kind of retraction associated to each paper is. This is accomplished by associating a score to each one of the reasons in the "Reason for Retraction" column, which will range from 0 to 4. The notebook then drops all papers with a severity score below a certain threshold (which is by default set to 3) so as to keep only those items that represent a genuine case of malpractice. 

The Notebook takes as input the clean data set that we generated in **Notebook 1a**. Its output, in turn, will be fed to **Notebook 1c.**

The **workflow** is therefore as follows:

- Input: **one .csv file** with our clean data set and **another .csv file** with the severity score correspondences that we created manually. 
- Output: **one .csv file** with the our clean data set, minus those papers with a severity score below the chosen threshhold. 


## Input / Output Parameters

Input parameters:


# Input path for severity score correspondence

input_path_severity_score = "../data/severity_score_assignements.csv"

# Input path for clean data set

input_path_data_set = "../data/retraction_watch_data_set/2_clean_data_set.csv"

# Severity threshold
# Only papers with this severity score or higher will be kept in our data frame

severity_threshold = 3

Output parameters:


# Output path for data set with severity score column

output_path = "../data/retraction_watch_data_set/3_severe_only_data_set.csv"

## Importing Libraries


# Import required libraries

import pandas as pd
import numpy as np
import function_definitions

## Adding a "Severity Score" to our Data Set

Let us start by loading our one-to-one correspondence between reasons for retraction and their associated severity score. We will do that by reading it from the .csv file that we created manually:


# Load severity score data from .csv file into data frame

df_severity = pd.read_csv(input_path_severity_score, sep = ";")

# Make sure column names are in snake case

df_severity.columns = [col.lower().replace(" ", "_") for col in df_severity.columns]

# Visualize result

df_severity

Next we will load the clean data set that we generated in Notebook 1:


# Read clean data set from input path

df = pd.read_csv(input_path_data_set, encoding='latin-1')

# Visualize result

df.head(1)

Having done that, we can go ahead and start the process to assign a severity score to each paper. Our first step will consist of adding a new column to our data frame, which will contain a list with all the reasons for retraction mentioned for that paper. We will first define a function that returns a list of individual reasons given a overall string of reasons separated by a given delimiter (such as the ones in the rows of our data frame, which are separated by a semicolon):


def separate_and_clean_reasons(reasons_string, delimiter = ";", extra_symbol = ""):
    """
    Function takes a single string containing multiple reasons separated by a given delimiter 
    and returns a list with all individual reasons as its elements, removing white spaces
    and any other extraneous symbols of choice, and the delimiter simbol.
    """
    
    # Return empty list is reasons string is NaN
    
    if pd.isna(reasons_string):
        return []

    # Store invididual reasons as separated by delimiter in list
    
    individual_reasons_list = reasons_string.split(delimiter)  
    
    # Intialize another list to store cleaned individual reasons
    
    clean_reasons_list = [] 
    
    # For loop to clean individual reasons by dropping delimiter and extra symbols
    
    for individual_reason in individual_reasons_list:
        
        # Clean individual reason and store it in new variable
        
        clean_individual_reason = individual_reason.strip(f"{delimiter}{extra_symbol} ")
        
        # If not empty, append clean individual reason to clean individual reasons list
        
        if clean_individual_reason:  
            clean_reasons_list.append(clean_individual_reason)
            
    # Return list with individual clean reasons
    
    return clean_reasons_list



We can now call our function to effectively add that extra column to our data frame. We will have to use the so-called lambda function in order to be able to use the "apply" method while passing the required arguments to our function:


# Call "separate_and_clean" function to add a column to data frame with lists of individual reasons as rows

df['reason_list'] = df['reason'].apply(lambda x: function_definitions.separate_and_clean_reasons(x, delimiter=";", extra_symbol="+"))


Next we will create an additional function that parses the content of each list of reasons, then returns the highest severity score of those associated to those reasons: 


# Define function to calculate the maximum severity score of those in the reasons_list column

def get_max_severity(reason_list):
    
    """Function takes a list with reasons for retraction, checks what are the severity scores
    associated to each individual reason, then returns the maximum score in that collection."""
    
    # Create list with scores associated to reasons in input list
    
    scores = df_severity[df_severity['reason'].isin(reason_list)]['severity_score']
    
    # Return maximum vaue in the scores list
    
    return max(scores) if not scores.empty else None


We can finally call our new function to add yet another column to our data frame, which will contain the highest severity score of those scores associated to the reasons why the paper was retracted:

# Call function to create new "severity_score" column with maximum severity score for each paper

df['severity_score'] = df['reason_list'].apply(get_max_severity)

# Convert elements of "severity_score" column to int type

df['severity_score'] = df['severity_score'].astype(int)


It will be important to get a sense of what the distribution of papers in our original data set per severity score is: 


# Obtain value count of each severity score in our data frame

score_counts = df['severity_score'].value_counts()

# Print value counts

score_counts


Let us also find out what percentage of the total entries in our data set those numbers represent. We will use the percentage_finder function that we defined in Notebook 1 for this purpose, calling it appropriately from the python file that we imported above:

function_definitions.percentage_printer(df, "severity_score")

We see how many papers have a severity score of 3 or more, which is good because we are only interested in investigating the most severe cases of misconduct. Given these percentages, we can go ahead and safely drop all papers with a severity score below our threshold (which was set to be equal to 3 at the beginning of the notebook but can be changed to have different values if necessary):



# Filter the dataframe to inlcude only those above our severity threshold

df_severe = df[df['severity_score'] >= severity_threshold]



Finally, we can also drop the column "reasons_list" from our data set, which we added as an intermediate step and won't be needing again in the future:


# Drop "reasons_list" column

df_severe.drop(columns=['reason_list'], inplace=True)

# Verify result

df_severe


Before going on to save are new data frame, let us check how many papers there are in it, and what fraction of the papers in our original data set that amounts to:


# Print total number of papers in data set

print(f"Total number of papers in data set: {df_severe.shape[0]}")

# Print percentage of papers retained compare with initial data set

print(f"Percentage of original data set that represents: {round(df_severe.shape[0] *100 / df.shape[0],2)} %")


## Output

To conclude, we write the resulting data frame into a .csv file:


# Save data frame as .csv

df.to_csv(output_path, index=False)
