==> 01c_select_subject.ipynb <==
# 1c. Data Cleaning: Dividing our Data Set Based on the "Subject" Column


## Introduction

This Notebook generates **a data frame with all the retracted papers in our data set from a specific field from the "Subject" column**, which will be selected by hand. Note that, since the "subject" column often mentions more than one subject, a single paper may be associated to more than one scientific field. 

The Notebook takes as input the clean data set generated by **Notebook 1b**, which is already filtered by severity score. The output that it generates will in turn be used by **Notebook 2a** in order to fetch text data from retracted papers.

The **workflow** is as follows:
    
- Input parameters: **one .csv** file with the clean data set with retracted papers filtered by severity score. 
- Output parameters: as many **.csv files** as unique entries there are in the "subject" column of the input .csv file.

## Input / Output Parameters

Input parameters:


# File path for input .csv file

input_path = "../data/retraction_watch_data_set/3_severe_only_data_set.csv"

# Code for subject to restrict analysis to

subject = "(BLS) Biology - Cellular"


Output parameters:


# File path for output .csv

output_path = '../data/retraction_watch_data_set/4_cell_bio_data_set.csv'  


## Importing Libraries

# Import required libraries

import pandas as pd
import numpy as np
import function_definitions

## Selecting a Subject

Let us start by loading our data from our .csv file into a dataframe:


# Read data from .csv file 

df = pd.read_csv(input_path, encoding='latin-1')

# Visualize first row of data frame

df.head(1)


Let us now take a look at the "subject" column:

df.subject


As we can see, the different rows in this column often contain more than one subject separated by a semicolon. It will be useful to find out how many unique subjects there are in our data frame:


# Create series with all subjects by splitting and exploding entries in subject column

subjects_series = df['subject'].str.split(';').explode()

# Obtain list with unique subjects from series

unique_subjects_list = sorted(subjects_series.unique())

# Visualize unique subjects

unique_subjects_list



It will also be useful to get a sense of how many papers we have per subject:


# Get series with value counts of each unique subject

subject_counts = subjects_series.value_counts()

# Visualize value counts

subject_counts



In the future, we will be interested in studying papers that belong to a single area of study. In order to be able to do this, let us start by defining a function that restricts our data set to a single one of all the subjects present in our data set:




def subject_selector(df, subject):
    """
    Function takes a data frame and a string with a subject name as input.
    It processes the 'subject' column by splitting its values separated by semicolons, 
    filters the data frame to include only rows with the specified subject, then
    returns the filtered data frame.
    """

    # Create series with exploded subjects by splitting the 'subject' column
    
    exploded_subjects = df['subject'].str.split(';').explode()
    
    # Drop the original 'subject' column and join the exploded subjects
    
    df = df.drop('subject', axis=1).join(exploded_subjects.rename('subject'))
    
    # Filter the data frame for the specified subject
    
    subject_df = df[df['subject'] == subject]
    
    # Return filtered data frame
    
    return subject_df

    


Since it is the subject with the most entries, molecular biology will be a good test subject case for our study. Let us thus call our function to restrict our data frame to papers in these area only (recall that the value of the variable "df_subject" had already been set that way when defining our input parameters above):


# Call function to create one .csv file with all retracted papers per subject

df_subject = subject_selector(df, subject)

# Check that it worked

df_subject


As always, it will be useful to get an idea of how many papers we have in our filtered data set, and of what percentage of the data set that we loaded at the beginning of this notebook that represents:


# Print total number of papers in data set

print(f"Total number of papers in data set: {df_subject.shape[0]}")

# Print percentage of papers retained compare with initial data set

print(f"Percentage of original data set that represents: {round(df_subject.shape[0] *100 / df.shape[0],2)} %")


## Output


Having done all this, we can now go ahead safe our new data frame, which has been restricted to a single subject, into a .csv file:


# Save date frame as .csv

df_subject.to_csv(output_path, index=False)
